requirements.txt
discord-py-slash-command
discord.py
langdetect
requests
translate
youtube-dl